+++ CHANGELOG.md
@@ -14,6 +14,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.8] - 2025-11-23
+### Added
+- Added link exchange program for paid users
+
+### Changed
+- using gtpr for writing posts
+
+### Fixed
+- correct link on project scan
+
 ## [0.0.8] - 2025-11-12
 ### Changed
 - Navbar spacing rules